+++ phoenix_example/lib/spec_example_web/endpoint.ex
@@ -1,0 +1,5 @@
+defmodule SpecExampleWeb.Endpoint do
+  use Phoenix.Endpoint, otp_app: :spec_example
+
+  socket "/socket", SpecExampleWeb.UserSocket
+end